requirements.txt
streamlit==1.15.2
pandas
numpy
seaborn
matplotlib
plotly
sklearn
lightgbm
shap
streamlit_shap
Pillow
xplotter